remove unnecessary packages, organize helpers, refactor
