+++ extendedQueryNodeSchema.graphql
