+++ src/helpers/data.ts
@@ -1,0 +1,12 @@
+type HasId = {
+  id: string
+}
+
+export const createLookup = <T extends HasId>(data: T[]): Record<string, T> => {
+  return data.reduce((acc, item) => {
+    if (item) {
+      acc[item.id] = item
+    }
+    return acc
+  }, {} as Record<string, T>)
+}
